Fixed experience-controller and restaurant-controller bugs

+++ backend/controllers/experience-controller.js
@@ -33,31 +33,31 @@ const createExperience = async (req, res, next) => {
 
 
 const updateAllExperiences = async (req, res, next) => {
-    try {
-        const restaurantId = req.body.restaurantId
-        console.log(`Received the following restaurantId: ${restaurantId}`);
+	try {
+		const restaurantId = req.body.restaurantId
+		console.log(`Received the following restaurantId: ${restaurantId}`);
 
-        const experiences = req.body.updatedExperienceData
-        const arrayLength = experiences.length
-        
-        // Loop through all experiences and update
+		const experiences = req.body.updatedExperienceData
+		const arrayLength = experiences.length
+
+		// Loop through all experiences and update
 		let i = 0
 		for (i = 0; i < arrayLength; i++) {
-            let experience = experiences[i]
+			let experience = experiences[i]
 
-            const experienceId = experience._id
-            delete experience._id
-            console.log(`Received the following experienceId: ${experienceId}`);
+			const experienceId = experience._id
+			delete experience._id
+			console.log(`Received the following experienceId: ${experienceId}`);
 
-            await Experience.updateOne(
-                {_id: experienceId},
-                {$set: experience}
-            ).lean()
+			await Experience.updateOne(
+				{ _id: experienceId },
+				{ $set: experience }
+			).lean()
 
-            let updatedExperience = await Experience.findById(experienceId)
-            console.log(updatedExperience)
+			let updatedExperience = await Experience.findById(experienceId)
+			console.log(updatedExperience)
 		}
-        return res.send(`Done.  Experiences of restaurantId: ${restaurantId} has been updated.`);
+		return res.send(`Done.  Experiences of restaurantId: ${restaurantId} has been updated.`);
 	} catch (err) {
 		return next(err);
 	}
@@ -65,8 +65,8 @@ const updateAllExperiences = async (req, res, next) => {
 
 
 const updateExperience = async (req, res, next) => {
-    try {
-        const restaurantId = req.body.restaurantId
+	try {
+		const restaurantId = req.body.restaurantId
 		delete req.body.restaurantId
 
 		const experienceId = req.body.experienceId
@@ -75,14 +75,14 @@ const updateExperience = async (req, res, next) => {
 		console.log(`Received the following restaurantId: ${restaurantId} and experienceId: ${experienceId}`);
 
 		await Experience.updateOne(
-			{_id: experienceId},
-			{$set: req.body}
-		).lean()    
-       
+			{ _id: experienceId },
+			{ $set: req.body }
+		).lean()
+
 		let updatedExperience = await Experience.findById(experienceId)
 		console.log(updatedExperience)
-		
-        return res.send(`Done.  The experienceId ${experienceId} of restaurantId: ${restaurantId} has been updated.`);
+
+		return res.send(`Done.  The experienceId ${experienceId} of restaurantId: ${restaurantId} has been updated.`);
 	} catch (err) {
 		return next(err);
 	}
@@ -93,7 +93,7 @@ const deleteExperience = async (req, res, next) => {
 	try {
 		const experienceId = req.body.experienceId
 		const countBefore = await Experience.countDocuments({ _id: experienceId })
-		await Experience.deleteOne({_id: experienceId})
+		await Experience.deleteOne({ _id: experienceId })
 		const countAfter = await Experience.countDocuments({ _id: experienceId })
 
 		return res.send(`Number of documents with experienceId: ${experienceId} is [Before Deletion: ${countBefore}] and [After Deletion: ${countAfter}].`);
@@ -107,14 +107,24 @@ const deleteExperience = async (req, res, next) => {
 const deleteAllExperiencesInteract = async (restaurantId) => {
 	try {
 		const restaurant = await Restaurant.findOne({ _id: restaurantId }).populate('experiences');
+
+		if (restaurant === null) {
+			return 0
+		}
+
 		const experiences = restaurant['experiences'];
+
+		if (experiences === null) {
+			return 0
+		}
+
 		const arrayLength = experiences.length;
 
 		// Loop through each of restaurant's experiences and delete
 		let i = 0
 		for (i = 0; i < arrayLength; i++) {
-			const experienceId = experiences[i]._id		
-			await Experience.deleteOne({_id: experienceId})
+			const experienceId = experiences[i]._id
+			await Experience.deleteOne({ _id: experienceId })
 		}
 
 		await Restaurant.updateOne(
@@ -147,11 +157,11 @@ const deleteAllExperiences = async (req, res, next) => {
 
 
 module.exports = {
-    getAllExperiences,
-    createExperience,
-    updateAllExperiences,
+	getAllExperiences,
+	createExperience,
+	updateAllExperiences,
 	updateExperience,
-    deleteExperience,
-    deleteAllExperiences,
+	deleteExperience,
+	deleteAllExperiences,
 	deleteAllExperiencesInteract
 }
